Update bybit sendorder

+++ vnpy/gateway/bybit/bybit_gateway.py
@@ -267,7 +267,7 @@ class BybitRestApi(RestClient):
         data = {
             "symbol": req.symbol,
             "side": DIRECTION_VT2BYBIT[req.direction],
-            "qty": int(req.volume),
+            "qty": float(req.volume),
             "order_link_id": orderid,
             "time_in_force": "GoodTillCancel",
             "reduce_only": False,
